requirements.txt
streamlit
sqlalchemy
passlib
python-dotenv
ollama
pymysql
mysql-connector-python
bcrypt
streamlit-google-auth
langchain
langchain-community
langchain-experimental
faiss-cpu
pdfplumber
torch
transformers
sentence-transformers
google-generativeai
langchain-google-genai